+++ src/server/shared/Packets/ByteBuffer.h
@@ -215,14 +215,14 @@ public:
 
     template <typename T> void WriteBits(T value, size_t bits)
     {
-        for (int32 i = bits - 1; i >= 0; --i)
+        for (size_t i = bits - 1; i >= 0; --i)
             WriteBit((value >> i) & 1);
     }
 
     uint32 ReadBits(size_t bits)
     {
         uint32 value = 0;
-        for (int32 i = bits - 1; i >= 0; --i)
+        for (size_t i = bits - 1; i >= 0; --i)
             if (ReadBit())
                 value |= (1 << (i));
